Changed float to double

+++ src/com/AILab3/Entities/Genes/ParetoGene.java
@@ -7,14 +7,14 @@ import static com.AILab3.GeneticAlgo.Constants.r;
 public class ParetoGene extends Gene
 {
     private static int min, max;
-    public float[] gene;
+    public double[] gene;
 
-    public ParetoGene (float[] _g)
+    public ParetoGene (double[] _g)
     {
         this(_g, 0, 0, 0);
     }
 
-    public ParetoGene (float[] _g, int _f, int _a, int _if)
+    public ParetoGene (double[] _g, int _f, int _a, int _if)
     {
         super(_f, _a, _if);
         gene = _g;
@@ -29,8 +29,8 @@ public class ParetoGene extends Gene
     @Override
     public int similar (Gene other)
     {
-        final float[] o = ((ParetoGene) other).gene;
-        final float[] t = this.gene;
+        final double[] o = ((ParetoGene) other).gene;
+        final double[] t = this.gene;
         int res = 0;
         for (int i = 0; i < this.gene.length; i++)
             if (o[i] == t[i]) res++;
@@ -60,7 +60,7 @@ public class ParetoGene extends Gene
     {
         final int l = this.getProblemSize();
         final float segmentSize = ((float) Math.abs(max - min)) / ((float) l), midSegment = segmentSize / 2;
-        float[] gen = new float[l];
+        double[] gen = new double[l];
         for (int j = 0; j < l; j++)
             gen[j] = r.nextBoolean() ? j * segmentSize + midSegment + r.nextFloat() * midSegment : j * segmentSize + midSegment - r.nextFloat() * midSegment;
         this.gene = gen;
